requirements.txt
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
Deprecated==1.2.13
idna==3.3
pycparser==2.21
PyGithub==1.55
PyJWT==2.4.0
PyNaCl==1.5.0
python-gitlab==3.4.0
requests==2.27.1
requests-toolbelt==0.9.1
urllib3==1.26.9
wrapt==1.14.1
